refactor: deduplicate buildResumeIngestData helper family into shared module
Extract 7 duplicated helper functions (toStringValue, toOptionalNumber,
toStringArray, parseBrandHits, parseRoleSignals, buildResumeIngestData,
isRecord) from resumes.ts, resumes_import.ts, and resumes_search.ts
into a new apps/api/src/services/resume-ingest-utils.ts shared module.

This eliminates 620 lines of triplicated code (-620/+247 net) and
fixes a latent bug in resumes_search.ts where isRecord was missing
the !Array.isArray(value) check, allowing arrays to pass the type
guard as Record<string, unknown>.
