+++ apps/api/src/routes/resumes_search.ts
@@ -10,7 +10,7 @@ import { AIMatchingService, type MatchingRequest, type MatchingResult } from "..
 import { MatchStorage, type MatchRunMode, type StoredMatch, type StoredMatchRun } from "../services/match-storage.js";
 import { SessionManager } from "../services/session-manager.js";
 import { JobDescriptionService } from "../services/job-description-service.js";
-import { RuleScoringService, type BrandContext, type BrandHit, type BrandRole, type RoleSignalSummary, type RuleScoringContext, type RuleScoringResult } from "../services/rule-scoring.js";
+import { RuleScoringService, type BrandHit, type RoleSignalSummary, type RuleScoringContext, type RuleScoringResult } from "../services/rule-scoring.js";
 import { SearchEventLogger } from "../services/search-event-logger.js";
 import { config } from "../services/config.js";
 import { DataNotFoundError } from "../services/errors.js";
@@ -46,6 +46,14 @@ import { SkillsKnowledgeService } from "../services/skills-knowledge.js";
 import type { ResumeItem } from "../types/resume.js";
 import type { ResumeIndex } from "../services/resume-index.js";
 import { workspaceConfigService } from "../services/workspace-config-service.js";
+import {
+  buildResumeIngestData,
+  parseBrandHits,
+  parseRoleSignals,
+  toOptionalNumber,
+  toStringArray,
+  toStringValue,
+} from "../services/resume-ingest-utils.js";
 
 const app = new OpenAPIHono();
 const resumeService = new ResumeService(config.projectRoot);
@@ -247,211 +255,6 @@ function resolveResumeSortOrder(
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null;
-}
-
-function toStringValue(value: unknown): string {
-  if (typeof value === "string") {
-    return value.trim();
-  }
-  if (value === null || value === undefined) {
-    return "";
-  }
-  return String(value).trim();
-}
-
-function toOptionalNumber(value: unknown): number | undefined {
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value !== "string") {
-    return undefined;
-  }
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  const parsed = Number(trimmed);
-  return Number.isFinite(parsed) ? parsed : undefined;
-}
-
-function toStringArray(value: unknown): string[] {
-  if (!Array.isArray(value)) {
-    return [];
-  }
-  return value
-    .map((item) => toStringValue(item))
-    .filter(Boolean);
-}
-
-function toBrandRole(value: unknown): BrandRole | null {
-  if (value === "employer" || value === "equipment" || value === "both") {
-    return value;
-  }
-  return null;
-}
-
-function toBrandContext(value: unknown): BrandContext | null {
-  if (
-    value === "employer"
-    || value === "equipment"
-    || value === "sales"
-    || value === "technical"
-    || value === "general"
-  ) {
-    return value;
-  }
-  return null;
-}
-
-function parseBrandHits(value: unknown): BrandHit[] {
-  if (!Array.isArray(value)) {
-    return [];
-  }
-
-  return value.flatMap((item) => {
-    if (!isRecord(item)) {
-      return [];
-    }
-
-    const brand = toStringValue(item.brand);
-    const role = toBrandRole(item.role);
-    const source = item.source === "workHistory" || item.source === "selfIntro" || item.source === "jobIntention"
-      ? item.source
-      : null;
-    const context = toBrandContext(item.context);
-
-    if (!brand || !role || !source || !context) {
-      return [];
-    }
-
-    return [{
-      brand,
-      role,
-      source,
-      context,
-    }];
-  });
-}
-
-function parseRoleSignals(value: unknown): RoleSignalSummary[] {
-  if (!Array.isArray(value)) {
-    return [];
-  }
-
-  return value.flatMap((item) => {
-    if (!isRecord(item)) {
-      return [];
-    }
-
-    const type = toStringValue(item.type);
-    const years = toOptionalNumber(item.years);
-    if (!type || years === undefined) {
-      return [];
-    }
-
-    const matchedSignals = toStringArray(item.matchedSignals);
-    const verifyIn = item.verifyIn === "searchText" ? "searchText" : "workHistory";
-    const signalCount = toOptionalNumber(item.signalCount) ?? matchedSignals.length;
-    const occurrences = toOptionalNumber(item.occurrences) ?? matchedSignals.length;
-    const industryVerifiedYears = toOptionalNumber(item.industryVerifiedYears) ?? 0;
-    const roleRelevantYears = toOptionalNumber(item.roleRelevantYears);
-    const industryVerifiedRelevantYears = toOptionalNumber(item.industryVerifiedRelevantYears);
-    const matchedWorkEntries = Array.isArray(item.matchedWorkEntries)
-      ? item.matchedWorkEntries.flatMap((entry) => {
-          if (!isRecord(entry)) {
-            return [];
-          }
-          const entryYears = toOptionalNumber(entry.years);
-          if (entryYears === undefined) {
-            return [];
-          }
-          return [{
-            companyName: toStringValue(entry.companyName) || undefined,
-            jobTitle: toStringValue(entry.jobTitle) || undefined,
-            years: entryYears,
-            industryVerified: entry.industryVerified === true,
-            matchedSignals: toStringArray(entry.matchedSignals),
-          }];
-        })
-      : undefined;
-
-    return [{
-      type,
-      matchedSignals,
-      signalCount,
-      occurrences,
-      years,
-      industryVerifiedYears,
-      ...(roleRelevantYears === undefined ? {} : { roleRelevantYears }),
-      ...(industryVerifiedRelevantYears === undefined ? {} : { industryVerifiedRelevantYears }),
-      ...(matchedWorkEntries && matchedWorkEntries.length > 0 ? { matchedWorkEntries } : {}),
-      verifyIn,
-    }];
-  });
-}
-
-function buildResumeIngestData(value: unknown): ResumeItem["ingestData"] | undefined {
-  if (!isRecord(value)) {
-    return undefined;
-  }
-
-  const industryTags = toStringArray(value.industryTags);
-  const synonymHits = toStringArray(value.synonymHits);
-  const evidenceText = toStringValue(value.evidenceText) || undefined;
-  const companyHits = toStringArray(value.companyHits);
-  const brandHits = parseBrandHits(value.brandHits);
-  const roleSignals = parseRoleSignals(value.roleSignals);
-  const industryDbV2Raw = toOptionalNumber(value.industryDbV2Raw);
-  const experienceLevel = toStringValue(value.experienceLevel) || undefined;
-  const normalizedExperienceLevel = experienceLevel?.trim().toLowerCase();
-  const meaningfulExperienceLevel = normalizedExperienceLevel && normalizedExperienceLevel !== 'unknown' ? experienceLevel : undefined;
-  const market = toStringValue(value.market) || undefined;
-  const ruleScores = isRecord(value.ruleScores)
-    ? Object.fromEntries(
-        Object.entries(value.ruleScores)
-          .filter((entry): entry is [string, number] => typeof entry[1] === 'number' && Number.isFinite(entry[1])),
-      )
-    : undefined;
-  const computedAt = toOptionalNumber(value.computedAt);
-  const skillsVersion = toOptionalNumber(value.skillsVersion);
-
-  if (
-    industryTags.length === 0
-    && synonymHits.length === 0
-    && !evidenceText
-    && companyHits.length === 0
-    && brandHits.length === 0
-    && roleSignals.length === 0
-    && industryDbV2Raw === undefined
-    && !meaningfulExperienceLevel
-    && !market
-    && (!ruleScores || Object.keys(ruleScores).length === 0)
-  ) {
-    return undefined;
-  }
-
-  const verifiedRoleYears = isRecord(value.verifiedRoleYears)
-    ? Object.fromEntries(
-        Object.entries(value.verifiedRoleYears)
-          .filter((entry): entry is [string, number] => typeof entry[1] === 'number' && Number.isFinite(entry[1])),
-      )
-    : undefined;
-
-  return {
-    ...(industryTags.length > 0 ? { industryTags } : {}),
-    ...(synonymHits.length > 0 ? { synonymHits } : {}),
-    ...(evidenceText ? { evidenceText } : {}),
-    ...(companyHits.length > 0 ? { companyHits } : {}),
-    ...(brandHits.length > 0 ? { brandHits } : {}),
-    ...(roleSignals.length > 0 ? { roleSignals } : {}),
-    ...(industryDbV2Raw === undefined ? {} : { industryDbV2Raw }),
-    ...(meaningfulExperienceLevel ? { experienceLevel: meaningfulExperienceLevel } : {}),
-    ...(verifiedRoleYears && Object.keys(verifiedRoleYears).length > 0 ? { verifiedRoleYears } : {}),
-    ...(ruleScores && Object.keys(ruleScores).length > 0 ? { ruleScores } : {}),
-    ...(market ? { market } : {}),
-    ...(computedAt !== undefined ? { computedAt } : {}),
-    ...(skillsVersion !== undefined ? { skillsVersion } : {}),
-  };
 }
 
 function parseConvexProvenance(value: unknown): ResumeSearchProvenance[] | undefined {
